update on task routes

+++ src/routers/tasks.js
@@ -1,12 +1,17 @@
 const express = require('express')
 const { update } = require('../models/task')
 const Task = require("../models/task")
+const auth = require("../middleware/Auth")
 
 const router = new express.Router()
 
 
-router.post('/tasks', async(req, res) => {
-    const task = new Task(req.body)
+router.post('/tasks', auth, async(req, res) => {
+
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    })
 
     try {
         await task.save()
@@ -14,14 +19,13 @@ router.post('/tasks', async(req, res) => {
     } catch (e) {
         res.status(400).send()
     }
-
-
 })
 
-router.get('/tasks', async(req, res) => {
+router.get('/tasks', auth, async(req, res) => {
 
     try {
-        const tasks = await Task.find({})
+        const tasks = await Task.find({ owner: req.user._id })
+            //  await req.user.populate('tasks').execPopulate()
         res.send(tasks)
     } catch (e) {
         res.status(500).send()
@@ -29,10 +33,11 @@ router.get('/tasks', async(req, res) => {
 
 })
 
-router.get('/tasks/:id', async(req, res) => {
+router.get('/tasks/:id', auth, async(req, res) => {
     const _id = req.params.id
     try {
-        const task = await Task.findById(_id)
+
+        const task = await Task.findOne({ _id, owner: req.user._id })
 
         if (!task) {
             return res.status(404).send()
@@ -80,10 +85,11 @@ router.patch('/tasks/:id', async(req, res) => {
 })
 
 
-router.delete('/tasks/:id', async(req, res) => {
+router.delete('/tasks/:id', auth, async(req, res) => {
 
     try {
-        const task = await Task.findByIdAndDelete(req.params.id)
+        //const task = await Task.findByIdAndDelete(req.params.id)
+        const task = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id })
         if (!task) {
             return res.status(400).send()
         }
